% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_ordpar.R
\name{bin_ordpar}
\alias{bin_ordpar}
\title{Bin Primary Metric Data into Histogram Bins and Compute Secondary Metric}
\usage{
bin_ordpar(df, minval, maxval, increment)
}
\arguments{
\item{df}{A data frame comsisting of a primary metric to be binned and a secondary to be computed}

\item{minval}{The minimum value of the primary metric determined via summary metrics}

\item{maxval}{The maximum value of the primary metric determined via summary metrics}

\item{increment}{The width/size of the histogram bin}
}
\value{
A data frame that is nicely sorted into bins according to primary metric
}
\description{
This function can be used after the choose_lower function where the data
frame is ready to be disassembled and binned into fixed width containers
where the resulting data frame is a lower bound inclusive histogram that
can be easily organized and plotted
}
