% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_lower.R
\name{choose_lower}
\alias{choose_lower}
\title{Filter the Lower Coordinate in a System with Multiple Measurements}
\usage{
choose_lower(ctrdf1, ctrdf2, ordpar1, ordpar2)
}
\arguments{
\item{ctrdf1}{First data frame of the first metric type}

\item{ctrdf2}{Second data frame of the first metric type}

\item{ordpar1}{First data frame of the second metric type}

\item{ordpar2}{Second data frame of th second metric type}
}
\value{
A data frame of the filtered secondary metric with respect to the first
}
\description{
This function works on raw computed molecular systems where
there are several areas of interest measured for the same metric.
But, in order to avoid duplication of data, i.e. the system was
counted more than once as it contains several metrics, this function
picks out only one of those metrics that correpsonds to a secondary
metric to the users choosing
}
